% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsgWordcloud.R
\docType{package}
\name{gsgWordcloud}
\alias{gsgWordcloud}
\alias{gsgWordcloud-package}
\title{API for Wordclouds}
\description{
The gsgWordcloud package contains functions in several categories:
}
\section{Replacements}{
 \code{\link{punctReplace}}, \code{\link{spellReplace}}
}

\section{Frequencies}{
 \code{\link{freqForWordCloud}}
}

\section{Generators}{
 \code{\link{generateWordCloudData}}
}

