% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell-punct.R
\name{spellReplace}
\alias{spellReplace}
\title{Replace all misspelled words with either the suggested replacement or the specified replacement}
\usage{
spellReplace(txt, keepWords = NULL, replaceWords = NULL)
}
\arguments{
\item{txt}{a character or character vector to be spell checked}

\item{keepWords}{a character vector of words to not replace with the inferred best word}

\item{replaceWords}{a character vector of the same length as \code{keepWords} of the user-defined replacements for the given words}
}
\value{
A list with two fields: \code{text}, a character or character vector with only single spaces, and \code{replacements}, a dataframe documenting each word replaced and what it was replaced by.
}
\description{
Replace all misspelled words with either the suggested replacement or the specified replacement
}
\note{
This will sometimes remove some/all punctuation and other times may add punctuation.
}
\examples{
collapseWS("hello    world") # "hello world"
}
